requirements.txt
google-generativeai>=0.5.0
edge-tts>=6.1.9
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0
pymupdf>=1.24.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
yfinance>=0.2.40
